+++ oniguruma/enc/utf32_le.c
@@ -148,7 +148,7 @@ utf32le_is_mbc_ambiguous(OnigCaseFoldType flag, const UChar** pp, const UChar* e
 static UChar*
 utf32le_left_adjust_char_head(const UChar* start, const UChar* s)
 {
-  int rem;
+  long rem;
 
   if (s <= start) return (UChar* )s;
